requirements.txt
tgcrypto
pyrogram
shapely
asyncio
python-dotenv
peewee
